Move signTx() to ChainUtil.
